Added package.json versions for node and npm
I started with the "~" versions that we use in our deployments.  Since
"engines" only throws a warning and not an error on install if the
environment doesn't match, this will serve as documentation without
breaking anyone's install.

We can widen the version acceptance range when we find evidence that
other versions are compatible.

Also added a human-readable classifier for python's setup.py.